DROID-363 Fix. 400 Waves in Leasing in History

+++ app/src/main/java/com/wavesplatform/wallet/v2/data/model/remote/response/Transaction.kt
@@ -23,7 +23,7 @@ open class Lease(
         @SerializedName("timestamp") var timestamp: Long = 0,
         @SerializedName("signature") var signature: String = "",
         @SerializedName("version") var version: Int = 0,
-        @SerializedName("amount") var amount: Int = 0,
+        @SerializedName("amount") var amount: Long = 0,
         @SerializedName("recipient") var recipient: String = ""
 ) : RealmModel
